+++ application/server/build.gradle.kts
@@ -26,12 +26,11 @@ dependencies {
     implementation(project(":module:swagger"))
     implementation(project(":module:auth"))
     implementation(project(":module:event"))
-    implementation(project(":module:reader"))
-    implementation(project(":module:updater"))
+    implementation(project(":module:search"))
     implementation(project(":module:validation"))
     implementation(project(":module:util"))
     implementation(project(":module:persistence"))
-    implementation(project(":module:reader"))
+    implementation(project(":module:search"))
 
     implementation(project(":domain:user"))
     implementation(project(":domain:client"))
