+++ application/server/src/main/kotlin/io/github/siyual_park/application/server/controller/UserController.kt
@@ -6,10 +6,10 @@ import io.github.siyual_park.application.server.dto.response.UserInfo
 import io.github.siyual_park.mapper.MapperManager
 import io.github.siyual_park.mapper.map
 import io.github.siyual_park.persistence.loadOrFail
-import io.github.siyual_park.reader.filter.RHSFilterParserFactory
-import io.github.siyual_park.reader.pagination.OffsetPage
-import io.github.siyual_park.reader.pagination.OffsetPaginator
-import io.github.siyual_park.reader.sort.SortParserFactory
+import io.github.siyual_park.search.filter.RHSFilterParserFactory
+import io.github.siyual_park.search.pagination.OffsetPage
+import io.github.siyual_park.search.pagination.OffsetPaginator
+import io.github.siyual_park.search.sort.SortParserFactory
 import io.github.siyual_park.user.domain.CreateUserPayload
 import io.github.siyual_park.user.domain.UserFactory
 import io.github.siyual_park.user.domain.UserStorage
