+++ domain/user/build.gradle.kts
@@ -13,7 +13,7 @@ dependencies {
     testImplementation(project(":module:data-test"))
     implementation(project(":module:auth"))
     implementation(project(":module:event"))
-    implementation(project(":module:reader"))
+    implementation(project(":module:search"))
     implementation(project(":module:persistence"))
     implementation(project(":module:mapper"))
