+++ module/search/src/main/kotlin/io/github/siyual_park/search/exception/FilterInvalidException.kt
@@ -1,0 +1,3 @@
+package io.github.siyual_park.search.exception
+
+class FilterInvalidException(message: String? = null) : InvalidRequestException(message)
