+++ module/search/src/main/kotlin/io/github/siyual_park/search/exception/InvalidRequestException.kt
@@ -1,0 +1,3 @@
+package io.github.siyual_park.search.exception
+
+open class InvalidRequestException(message: String? = null) : RuntimeException(message)
