+++ module/search/src/main/kotlin/io/github/siyual_park/search/exception/SortInvalidException.kt
@@ -1,0 +1,3 @@
+package io.github.siyual_park.search.exception
+
+class SortInvalidException(message: String? = null) : InvalidRequestException(message)
