+++ module/search/src/main/kotlin/io/github/siyual_park/search/pagination/OffsetPageResponseHandler.kt
@@ -1,0 +1,17 @@
+package io.github.siyual_park.search.pagination
+
+import io.github.siyual_park.viewer.ResponseBodyResultHandlerAdapter
+import org.springframework.http.codec.ServerCodecConfigurer
+import org.springframework.web.bind.annotation.ControllerAdvice
+import org.springframework.web.reactive.accept.RequestedContentTypeResolver
+
+@ControllerAdvice
+class OffsetPageResponseHandler(
+    serverCodecConfigurer: ServerCodecConfigurer,
+    resolver: RequestedContentTypeResolver,
+    presenter: OffsetPagePresenter
+) : ResponseBodyResultHandlerAdapter<OffsetPage<*>>(
+    serverCodecConfigurer.writers,
+    resolver,
+    presenter
+)
