--- module/updater/.gitignore
@@ -1,7 +1,0 @@
-/.gradle
-/.idea
-/out
-/build
-*.iml
-*.ipr
-*.iws
