--- module/updater/src/main/kotlin/io/github/siyual_park/updater/patch/JsonPatch.kt
@@ -1,13 +1,0 @@
-package io.github.siyual_park.updater.patch
-
-import com.fasterxml.jackson.databind.ObjectMapper
-import io.github.siyual_park.data.patch.Patch
-
-class JsonPatch<T, U>(
-    private val objectMapper: ObjectMapper,
-    private val overrides: U
-) : Patch<T> {
-    override fun apply(entity: T): T {
-        return objectMapper.updateValue(entity, overrides)
-    }
-}
