--- module/updater/src/main/kotlin/io/github/siyual_park/updater/patch/JsonPatchFactory.kt
@@ -1,11 +1,0 @@
-package io.github.siyual_park.updater.patch
-
-import com.fasterxml.jackson.databind.ObjectMapper
-import org.springframework.stereotype.Component
-
-@Component
-class JsonPatchFactory(
-    private val objectMapper: ObjectMapper,
-) {
-    fun <T, U> create(overrides: U) = JsonPatch<T, U>(objectMapper, overrides)
-}
